pyproject.toml
[tool.poetry]
name = "birbia-station-bot"
version = "2.0.0"
description = "Birbia's Radio discord bot."
authors = ["SkinnyDevi <skinnydevi.dev@gmail.com>"]
maintainers = ["SkinnyDevi <skinnydevi.dev@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]
repository = "https://github.com/SkinnyDevi/birbia-station-bot"

[tool.poetry.dependencies]
python = "^3.10.6"
yt-dlp = "^2023.7.6"
async-timeout = ">=4.0.0a3,<5.0"
PyNaCl = "^1.4.0"
websockets = "^9.1"
aiohttp = "^3.8.5"
"discord.py" = "^2.3.2"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
MarkupSafe = "2.0.1"
beautifulsoup4 = "^4.11.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
certifi = "^2023.07.22"
future = "^0.18.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pycodestyle = "^2.11.0"

[tool.poetry.scripts]
bot = "src.core.utils.poetry_utils:start_dev"
test = "src.core.utils.poetry_utils:start_tests"
build = "src.core.utils.poetry_utils:docker_build"
mount = "src.core.utils.poetry_utils:docker_make_container"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
